Better balance obscure side-effect ordering issues.
Recombine the loops in CreateDynamicFunction that build the
formal parameter list and find the bodyText into one.

This is necessary so that stringifier side-effects for parameters
happen before any side-effects due to coercion of the body text.

Also, moved the large, more explanatory changes to the host callout
definition above the changes to clients of the callout.